chore: start porting over validation operator methods
